Use a non-blocking interface to libvirt
Create a non-blocking interface to libvirt by running the libvirt
operations in a thread. The thread servicing Erlang and the thread
running libvirt operations communicate over a socket.

The Erlang end is non-blocking and will return immediately. If the libvirt
end is not reading quickly enough, the Erlang side should be returned
an error message (I'm assuming eagain, but I haven't tested this theory
yet!).

The Erlang side sends a struct that basically contains an ErlNifFunc.
This is used to call the actual blocking function that used to be exposed
via the NIF interface.

The libvirt thread stays in a loop, blocking on its socket. When data
is sent over the socket, the thread reads it and through the magic of
icky shared state, the libvirt wrapping function is called.

Tested by running the example scripts. Nothing segfaulted. I declare
success!